Refactor Stripe module into modular directory structure
Split lib/stripe.ts into lib/stripe/ directory:
- lib/stripe/index.ts - Server-only Stripe client and functions
- lib/stripe/plans.ts - Client-safe plan constants and types

This separation:
- Prevents server env vars from leaking to client bundles
- Allows safe import of plan constants on client side
- Provides proper getPriceIdForPlan() function for server-side price ID resolution
